requirements.txt
openai
langchain
python-dotenv
tiktoken
pinecone-client
streamlit